refactor: array file structure + AttributeField -> AttributeFieldSelector
