+++ packages/dm-core-plugins/src/form/fields/ArrayField/templates/index.ts
@@ -1,6 +1,6 @@
 import { TTemplates } from '../../../types'
 import ArrayPrimitiveDatagridTemplate from './ArrayPrimitiveDataGridTemplate'
-import ArrayPrimitiveListTemplate from './ArrayPrimitiveListTemplate'
+import ArrayPrimitiveListTemplate from './ArrayPrimitiveListTemplate/ArrayPrimitiveListTemplate'
 
 const arrayTemplates: TTemplates = {
   ArrayPrimitiveDatagridTemplate,
